progress listener v2

+++ k4l-video-trimmer/src/main/java/life/knowledge4/videotrimmer/interfaces/OnTrimVideoListener.java
@@ -35,5 +35,5 @@ public interface OnTrimVideoListener {
 
     void onError(final String message);
 
-    void onProgress(int progress);
+    void onProgress(float progress);
 }
